Rename UpgradeTest to UpgradesMatrix, split Unit test into own file
