+++ sdk/daml-lf/engine/BUILD.bazel
@@ -88,7 +88,8 @@ da_scala_test_suite(
         exclude = [
             "src/test/**/LargeTransactionTest.scala",
             "src/test/**/MinVersionTest.scala",
-            "src/test/**/UpgradeTest.scala",
+            "src/test/**/UpgradesMatrix.scala",
+            "src/test/**/UpgradesMatrixUnit.scala",
         ],
     ),
     data = [
@@ -204,7 +205,8 @@ da_scala_test(
     name = "upgrade-test",
     timeout = "moderate",
     srcs = glob([
-        "src/test/**/UpgradeTest.scala",
+        "src/test/**/UpgradesMatrix.scala",
+        "src/test/**/UpgradesMatrixUnit.scala",
     ]),
     args = ["-P"],  # run suites in parallel
     scala_deps = [
